Add files via upload
Changed some spelling, changed the scientific notation to round to two decimal points (Graphs are easier to read prices a lot easier in my opinion), and separated the Ethereum and XRP graphs (XRP doesn't even get over $2 so it looks like a steady $0 compared to Ethereum and Bitcoin. Would be a different story if we were looking at market capitalization though).